refactor LocaDB to SpotData, PlaceData and renamed Reporter to NavioLogger
